pyproject.toml
[project]
name = "kaizen"
version = "0.1.0"
description = "Kaizen — continuous prompt optimization platform for AI teams"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "alembic==1.18.4",
    "celery==5.6.3",
    "dspy==3.1.3",
    "fastapi==0.135.1",
    "greenlet>=3.3.2",
    "litellm==1.80.12",
    "psycopg[binary]==3.3.3",
    "pydantic-settings==2.13.1",
    "redis>=7.4.0",
    "sqlalchemy==2.0.48",
    "uvicorn==0.42.0",
    "PyGithub==2.3.0",
    "cryptography==42.0.5",
    "python-multipart>=0.0.22",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
